Re-export indexer fuel-indexer-* dependencies under fuel-indexer-utils
