+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
    "examples/hello-world/hello-indexer",
    "examples/hello-world/hello-world-data",
    "examples/hello-world/hello-world-node",
+   "packages/fuel-indexer",
    "packages/fuel-indexer-api-server",
    "packages/fuel-indexer-database",
    "packages/fuel-indexer-database/database-types",
@@ -24,8 +25,7 @@ members = [
    "packages/fuel-indexer-tests/components/indices/simple-wasm/simple-wasm",
    "packages/fuel-indexer-tests/components/web-api",
    "packages/fuel-indexer-types",
-   "packages/fuel-indexer",
-   "plugins/forc-index-tests",
+   "packages/fuel-indexer-utils",
    "plugins/forc-index",
    "plugins/forc-index-tests",
    "plugins/forc-postgres",
@@ -43,6 +43,7 @@ default-members = [
    "packages/fuel-indexer-plugin",
    "packages/fuel-indexer-schema",
    "packages/fuel-indexer-types",
+   "packages/fuel-indexer-utils",
    "plugins/forc-index",
    "plugins/forc-postgres",
 ]
@@ -77,6 +78,7 @@ fuel-indexer-plugin = { version = "0.15.1", path = "./packages/fuel-indexer-plug
 fuel-indexer-postgres = { version = "0.15.1", path = "./packages/fuel-indexer-database/postgres" }
 fuel-indexer-schema = { version = "0.15.1", path = "./packages/fuel-indexer-schema", default-features = false }
 fuel-indexer-types = { version = "0.15.1", path = "./packages/fuel-indexer-types" }
+fuel-indexer-utils = { version = "0.15.1", path = "./packages/fuel-indexer-utils" }
 clap = "3.1"
 bincode = "1.3.3"
 fuel-tx = { version = "0.26", default-features = false }
