+++ examples/hello-world-native/hello-indexer-native/Cargo.toml
@@ -7,8 +7,6 @@ publish = false
 [dependencies]
 async-trait = { version = "0.1" }
 fuel-indexer = { workspace = true }
-fuel-indexer-macros = { workspace = true, default-features = false }
-fuel-indexer-plugin = { workspace = true, features = ["native-execution"] }
-fuel-indexer-schema = { workspace = true, default-features = false }
+fuel-indexer-utils = { workspace = true, features = ["native-execution"] }
 fuels = { workspace = true }
 serde = { workspace = true }
