+++ packages/fuel-indexer-macros/Cargo.toml
@@ -31,6 +31,7 @@ syn = { version = "2.0", features = ["full"] }
 
 [dev-dependencies]
 fuel-indexer-plugin = { workspace = true }
+fuel-indexer-utils = { workspace = true }
 fuels-macros = { version = "0.40.0", default-features = false }
 fuels-types = { version = "0.40.0", default-features = false }
 serde = { workspace = true }
