+++ packages/fuel-indexer-macros/src/native.rs
@@ -31,20 +31,26 @@ pub fn handler_block_native(
 /// indexer module, not within the scope of the entire lib module.
 fn native_prelude() -> proc_macro2::TokenStream {
     quote! {
-        use fuel_indexer_plugin::native::*;
-        use fuel_indexer_plugin::{{serialize, deserialize}, types::*, serde::{Deserialize, Serialize}, serde_json};
-
-        // TODO: Eventually prevent these types of implicity imports and have users import
-        // all dependencies explicity (preferably through a single crate).
-        use fuels::{
-            core::abi_decoder::ABIDecoder,
-            macros::{Parameterize, Tokenizable},
-            types::{StringToken, traits::{Tokenizable, Parameterize}},
-        };
-
         type B256 = [u8; 32];
 
         static mut db: Option<Arc<Mutex<Database>>> = None;
 
+        use fuel_indexer_utils::{
+            fuels::{
+                core::abi_decoder::ABIDecoder,
+                macros::{Parameterize, Tokenizable},
+                types::{
+                    traits::{Parameterize, Tokenizable},
+                    StringToken,
+                },
+            },
+            native::*,
+            prelude::{
+                bincode, deserialize,
+                serde_json, serialize,
+            },
+            serde::{Deserialize, Serialize},
+            types::*,
+        };
     }
 }
