+++ packages/fuel-indexer-macros/src/wasm.rs
@@ -12,7 +12,7 @@ pub fn handler_block_wasm(
 
         #[no_mangle]
         fn handle_events(blob: *mut u8, len: usize) {
-            use fuel_indexer_plugin::deserialize;
+            use fuel_indexer_utils::prelude::deserialize;
             let bytes = unsafe { Vec::from_raw_parts(blob, len, len) };
             let blocks: Vec<BlockData> = match deserialize(&bytes) {
                 Ok(blocks) => blocks,
@@ -38,25 +38,26 @@ pub fn handler_block_wasm(
 fn wasm_prelude() -> proc_macro2::TokenStream {
     quote! {
         use alloc::{format, vec, vec::Vec};
-        use fuel_indexer_plugin::{
-            serialize,
-            deserialize,
-            wasm::{Logger, FromHex, Digest, Sha256, Entity},
-            types::*,
-            bincode,
-            serde::{Deserialize, Serialize},
-            serde_json
-        };
-
-        // TODO: Eventually prevent these types of implicity imports and have users import
-        // all dependencies explicity (preferably through a single crate).
-        use fuels::{
-            core::abi_decoder::ABIDecoder,
-            macros::{Parameterize, Tokenizable},
-            types::{StringToken, traits::{Tokenizable, Parameterize}},
-        };
         use std::str::FromStr;
 
         type B256 = [u8; 32];
+
+        use fuel_indexer_utils::{
+            fuels::{
+                core::abi_decoder::ABIDecoder,
+                macros::{Parameterize, Tokenizable},
+                types::{
+                    traits::{Parameterize, Tokenizable},
+                    StringToken,
+                },
+            },
+            prelude::{
+                bincode, deserialize,
+                serde_json, serialize,
+            },
+            serde::{Deserialize, Serialize},
+            types::*,
+            wasm::{Digest, Entity, FromHex, Logger, Sha256},
+        };
     }
 }
