+++ packages/fuel-indexer-plugin/src/lib.rs
@@ -25,13 +25,7 @@ pub mod types {
 }
 
 pub mod utils {
-    pub use fuel_indexer_lib::utils::{
-        indexer_utils::{
-            bytes32_from_inputs, first32_bytes_to_bytes32, first8_bytes_to_u64,
-            trim_sized_ascii_string, u64_id, u64_id_from_inputs,
-        },
-        sha256_digest,
-    };
+    pub use fuel_indexer_lib::utils::sha256_digest;
 }
 
 pub use bincode;
