+++ packages/fuel-indexer-tests/components/indices/fuel-indexer-test/Cargo.toml
@@ -8,8 +8,6 @@ publish = false
 crate-type = ['cdylib']
 
 [dependencies]
-fuel-indexer-macros = { workspace = true, default-features = false }
-fuel-indexer-plugin = { workspace = true }
-fuel-indexer-schema = { workspace = true, default-features = false }
+fuel-indexer-utils = { workspace = true }
 fuels = { workspace = true }
 serde = { workspace = true }
