+++ packages/fuel-indexer-tests/components/indices/simple-wasm/simple-wasm/Cargo.toml
@@ -8,9 +8,7 @@ publish = false
 crate-type = ['cdylib']
 
 [dependencies]
-fuel-indexer-macros = { workspace = true, default-features = false }
-fuel-indexer-plugin = { workspace = true }
-fuel-indexer-schema = { workspace = true, default-features = false }
+fuel-indexer-utils = { workspace = true }
 fuel-tx = { workspace = true }
 fuels = { workspace = true }
 serde = { workspace = true }
