+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.stderr
@@ -1,4 +1,4 @@
-error: Type with ident 'Ident { ident: "BadType", span: #0 bytes(217..224) }' not defined in the ABI.
+error: Type with ident 'Ident { ident: "BadType", span: #0 bytes(224..231) }' not defined in the ABI.
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.rs
   |
   | #[indexer(manifest = "packages/fuel-indexer-tests/trybuild/simple_wasm.yaml")]
@@ -68,6 +68,8 @@ help: consider importing one of these items
    |
 2  | use fuel_indexer_types::fuel::BlockData;
    |
+2  | use fuel_indexer_utils::prelude::BlockData;
+   |
 
 error[E0422]: cannot find struct, variant or union type `BlockData` in this scope
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.rs
@@ -80,6 +82,8 @@ help: consider importing one of these items
 2  | use fuel_indexer_plugin::prelude::BlockData;
    |
 2  | use fuel_indexer_types::fuel::BlockData;
+   |
+2  | use fuel_indexer_utils::prelude::BlockData;
    |
 
 error[E0433]: failed to resolve: use of undeclared crate or module `fuel`
@@ -116,11 +120,53 @@ help: consider importing one of these items
    |
 2  | use fuel_indexer_types::fuel::Transaction;
    |
-2  | use fuels::prelude::Transaction;
+2  | use fuel_indexer_utils::prelude::HeaderData;
    |
-2  | use fuels::tx::Transaction;
+
+error[E0422]: cannot find struct, variant or union type `TransactionData` in this scope
+ --> ../fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.rs
+  |
+  |         transactions: vec![TransactionData {
+  |                            ^^^^^^^^^^^^^^^ not found in this scope
+  |
+help: consider importing one of these items
    |
-     and 1 other candidate
+2  | use fuel_indexer_plugin::prelude::TransactionData;
+   |
+2  | use fuel_indexer_types::prelude::TransactionData;
+   |
+2  | use fuel_indexer_utils::prelude::TransactionData;
+   |
+
+error[E0433]: failed to resolve: use of undeclared type `ClientTransactionStatusData`
+ --> ../fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.rs
+  |
+  |             status: ClientTransactionStatusData::default(),
+  |                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^ use of undeclared type `ClientTransactionStatusData`
+  |
+help: consider importing one of these items
+   |
+2  | use fuel_indexer_plugin::prelude::ClientTransactionStatusData;
+   |
+2  | use fuel_indexer_types::prelude::ClientTransactionStatusData;
+   |
+2  | use fuel_indexer_utils::prelude::ClientTransactionStatusData;
+   |
+
+error[E0433]: failed to resolve: use of undeclared type `ClientTransaction`
+ --> ../fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.rs
+  |
+  |             transaction: ClientTransaction::default(),
+  |                          ^^^^^^^^^^^^^^^^^ use of undeclared type `ClientTransaction`
+  |
+help: consider importing one of these items
+   |
+2  | use fuel_indexer_plugin::prelude::ClientTransaction;
+   |
+2  | use fuel_indexer_types::prelude::ClientTransaction;
+   |
+2  | use fuel_indexer_utils::prelude::ClientTransaction;
+   |
 
 error[E0425]: cannot find function `serialize` in this scope
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.rs
@@ -133,6 +179,8 @@ help: consider importing one of these items
 2  | use fuel_indexer_lib::utils::serialize;
    |
 2  | use fuel_indexer_plugin::serialize;
+   |
+2  | use fuel_indexer_utils::serialize;
    |
 
 error[E0425]: cannot find function `handle_events` in this scope
