+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_args_include_self.stderr
@@ -1,10 +1,8 @@
 error: `self` argument not allowed in handler function.
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_args_include_self.rs
   |
-  | / #[indexer(
-  | |     manifest = "packages/fuel-indexer-tests/trybuild/simple_wasm.yaml"
-  | | )]
-  | |__^
+  | #[indexer(manifest = "packages/fuel-indexer-tests/trybuild/simple_wasm.yaml")]
+  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
   = note: this error originates in the attribute macro `indexer` (in Nightly builds, run with -Z macro-backtrace for more info)
