+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_args_not_included.rs
@@ -1,4 +1,4 @@
-use fuel_indexer_macros::indexer;
+use fuel_indexer_utils::macros::indexer;
 
 #[indexer()]
 mod indexer {
