+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
@@ -1,5 +1,5 @@
 extern crate alloc;
-use fuel_indexer_macros::indexer;
+use fuel_indexer_utils::macros::indexer;
 
 #[no_mangle]
 fn ff_log_data(_inp: ()) {}
