+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.stderr
@@ -68,6 +68,8 @@ help: consider importing one of these items
    |
 2  | use fuel_indexer_types::fuel::BlockData;
    |
+2  | use fuel_indexer_utils::prelude::BlockData;
+   |
 
 error[E0422]: cannot find struct, variant or union type `BlockData` in this scope
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
@@ -81,6 +83,8 @@ help: consider importing one of these items
    |
 2  | use fuel_indexer_types::fuel::BlockData;
    |
+2  | use fuel_indexer_utils::prelude::BlockData;
+   |
 
 error[E0433]: failed to resolve: use of undeclared type `Consensus`
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
@@ -93,6 +97,8 @@ help: consider importing one of these items
 2  | use fuel_indexer_plugin::prelude::fuel::Consensus;
    |
 2  | use fuel_indexer_types::fuel::Consensus;
+   |
+2  | use fuel_indexer_utils::prelude::HeaderData;
    |
 
 error[E0433]: failed to resolve: use of undeclared crate or module `fuel`
@@ -129,11 +135,53 @@ help: consider importing one of these items
    |
 2  | use fuel_indexer_types::fuel::Transaction;
    |
-2  | use fuels::prelude::Transaction;
+2  | use fuel_indexer_utils::prelude::ConsensusData;
    |
-2  | use fuels::tx::Transaction;
+
+error[E0422]: cannot find struct, variant or union type `TransactionData` in this scope
+ --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
+  |
+  |         transactions: vec![TransactionData {
+  |                            ^^^^^^^^^^^^^^^ not found in this scope
+  |
+help: consider importing one of these items
    |
-     and 1 other candidate
+2  | use fuel_indexer_plugin::prelude::TransactionData;
+   |
+2  | use fuel_indexer_types::prelude::TransactionData;
+   |
+2  | use fuel_indexer_utils::prelude::TransactionData;
+   |
+
+error[E0433]: failed to resolve: use of undeclared type `ClientTransactionStatusData`
+ --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
+  |
+  |             status: ClientTransactionStatusData::default(),
+  |                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^ use of undeclared type `ClientTransactionStatusData`
+  |
+help: consider importing one of these items
+   |
+2  | use fuel_indexer_plugin::prelude::ClientTransactionStatusData;
+   |
+2  | use fuel_indexer_types::prelude::ClientTransactionStatusData;
+   |
+2  | use fuel_indexer_utils::prelude::ClientTransactionStatusData;
+   |
+
+error[E0433]: failed to resolve: use of undeclared type `ClientTransaction`
+ --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
+  |
+  |             transaction: ClientTransaction::default(),
+  |                          ^^^^^^^^^^^^^^^^^ use of undeclared type `ClientTransaction`
+  |
+help: consider importing one of these items
+   |
+2  | use fuel_indexer_plugin::prelude::ClientTransaction;
+   |
+2  | use fuel_indexer_types::prelude::ClientTransaction;
+   |
+2  | use fuel_indexer_utils::prelude::ClientTransaction;
+   |
 
 error[E0425]: cannot find function `serialize` in this scope
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
@@ -146,6 +194,8 @@ help: consider importing one of these items
 2  | use fuel_indexer_lib::utils::serialize;
    |
 2  | use fuel_indexer_plugin::serialize;
+   |
+2  | use fuel_indexer_utils::serialize;
    |
 
 error[E0425]: cannot find function `handle_events` in this scope
