+++ packages/fuel-indexer-tests/trybuild/pass_if_indexer_is_valid_single_type.rs
@@ -1,5 +1,5 @@
 extern crate alloc;
-use fuel_indexer_macros::indexer;
+use fuel_indexer_utils::macros::indexer;
 
 #[no_mangle]
 fn ff_log_data(_inp: ()) {}
