+++ packages/fuel-indexer-utils/src/lib.rs
@@ -1,0 +1,6 @@
+pub use fuel_indexer_macros as macros;
+pub use fuel_indexer_plugin::*;
+pub use fuel_indexer_schema;
+pub use fuels;
+
+pub mod utilities;
